Update to GAE 1.7.5
Add ability to check app version of submitted report against the most recent version in the mapping list, if the versions dont match we can discard the report and return "OLD REPORT".

Will create a ACRA submit which sees "OLD REPORT" and posts an intent to the application to tell the user to upgrade.

Problem here, i had originally set the mapping date to not indexed, so it has to be re-saved to create the index, the recalc totals script does this for the user..